refactor: only regenerate spread ranks when floor or species changes
there is no need to regenerate them every time our subject IVs change, so we can do the searching for the subject in a memoized list of ranks. this has huge performance benefits when checking different IV spreads for the same species.

this also allows us to have a "maximize" button which quickly populates the subject's IVs with the top-ranked IV spread for the requested cp/level cap