refactor: move calculation of rankedSpreads to LeagueContent-scoped context
This is a pretty significant performance increase. The data which is used for both LeagueSubjectAtLevel and LeagueTopSpreadsForLevel are the same, so we can extract the generation of the data to their parent and then make it available to them via Context.

This means the dependencies changing (level cap, subject species, subject floor) is the only way to make this calculation re-run. With the previous implementation the calculation would be run every time the LeagueSubjectAtLevel or LeagueTopSpreadsAtLevel were mounted, so switching between the two views is now significantly less expensive.